feat: remove social network icons
Site owner does not use social media for now.
Also the icons are not rendered locally by backstop; they are rendering
in the CI, thus CI tests fail as the reference images do not contain icons.
This might be a problem of local backstop setup, but we don't need to
investigate it for now.